requirements.txt
python-osc
python-dotenv